requirements.txt
# Django
# ------------------------------------------------------------------------------
django==3.2.4

# Restframework
djangorestframework==3.12.4

requests==2.25.1
